--- sdk/dir-js/example/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "dir-example",
-  "version": "0.0.0",
-  "scripts": {
-    "example": "node example.js"
-  },
-  "dependencies": {
-  }
-}
