+++ sdk/dir-js/models/core_v1.ts
@@ -1,4 +1,4 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '@buf/agntcy_dir.bufbuild_es/agntcy/dir/core/v1/record_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/core/v1/record_pb';
