+++ sdk/dir-js/models/store_v1.ts
@@ -1,5 +1,7 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '@buf/agntcy_dir.bufbuild_es/agntcy/dir/store/v1/store_service_pb';
-export * from '@buf/agntcy_dir.bufbuild_es/agntcy/dir/store/v1/sync_service_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/store/v1/store_service_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/store/v1/store_service_pb.client';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/store/v1/sync_service_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/store/v1/sync_service_pb.client';
