+++ sdk/examples/example-js/package.json
@@ -1,11 +1,9 @@
 {
   "name": "dir-example",
   "version": "0.0.0",
-  "type": "module",
   "scripts": {
     "example": "node example.js"
   },
   "dependencies": {
-    "agntcy-dir": "file:../../dir-js"
   }
 }
